sudarshan-ocr-apiserver/requirements.txt
fastapi
uvicorn
opencv-python-headless
pillow
pytesseract
matplotlib
numpy
mangum
# uvicorn app.main:app --reload
# http://127.0.0.1:8000/